Seperate astronomicals to different files
