+++ Cargo.toml
@@ -18,3 +18,4 @@ env_logger = "0.3"
 nalgebra = "0.13.1"
 statrs = "0.9.0"
 lazy_static = "1.0"
+derive_builder = "0.5.1"
