+++ src/main.rs
@@ -11,6 +11,8 @@ extern crate nalgebra;
 extern crate statrs;
 #[macro_use]
 extern crate lazy_static;
+#[macro_use]
+extern crate derive_builder;
 
 #[macro_use(info, debug, log)]
 extern crate log;
